+++ src/views/EventsView/index.tsx
@@ -1,26 +1,55 @@
 'use client';
-import EventList from '@/components/EventList';
+import React, {useState} from 'react';
+import { Box, Button, Link, Tab, Tabs, Typography } from '@mui/material';
 import { EventResponse } from '@/types/dataModel/event';
-import { Box, Button, Typography } from '@mui/material';
-import Link from 'next/link';
+import dayjs from 'dayjs';
+import EventList from '@/components/EventList';
 import SearchField from '@/components/SearchField';
 import useSearch from '@/hooks/useSearch';
 
 interface EventsViewProps {
-  events: EventResponse[];
+  events: EventResponse[]
 }
 
-export default function EventsView({ events }: EventsViewProps) {
-  const search = useSearch();
+const sortEventsByDateAsc = (events: EventResponse[]) => {
+  return events.sort((a, b) => {
+    return dayjs(a.date).valueOf() - dayjs(b.date).valueOf();
+  });
+};
 
-  if (search.length > 0) {
-    events = events.filter((event) =>
-      `${event.name}`.toLowerCase().includes(search.toLowerCase())
-    );
-  }
+const sortEventsByDateDesc = (events: EventResponse[]) => {
+  return events.sort((a, b) => {
+    return dayjs(b.date).valueOf() - dayjs(a.date).valueOf();
+  });
+};
 
-  return (
-    <Box>
+export default function EventsPage({events}: EventsViewProps) {
+    const search = useSearch();
+
+    if (search.length > 0) {
+      events = events.filter((event) =>
+        `${event.name}`.toLowerCase().includes(search.toLowerCase())
+      );
+    }
+
+    let upcomingEvents: EventResponse[];
+    let pastEvents: EventResponse[];
+
+    const today = dayjs().startOf('day').toDate();
+    const eventsBeforeToday = events.filter(event => dayjs(event.date).isBefore(today));
+    const eventsTodayAndAfter = events.filter(event => dayjs(event.date).isSame(today, 'day') || dayjs(event.date).isAfter(today));
+        
+    upcomingEvents = sortEventsByDateAsc(eventsTodayAndAfter);
+    pastEvents = sortEventsByDateDesc(eventsBeforeToday);
+
+    const [tabIndex, setTabIndex] = React.useState(0);
+    const handleTabChange = (event: React.SyntheticEvent, newValue: number) => {
+      setTabIndex(newValue);
+    };
+
+    return (
+      <>
+      <Box>
       <SearchField />
       <Typography variant="h3" sx={{ my: 2 }}>
         Events
@@ -30,7 +59,17 @@ export default function EventsView({ events }: EventsViewProps) {
           New event
         </Button>
       </Link>
-      <EventList events={events} />
+        <Tabs value={tabIndex} onChange={handleTabChange} centered>
+          <Tab label="Upcoming Events" />
+          <Tab label="Past Events" />
+        </Tabs>
+        {tabIndex === 0 && (
+          <EventList events={upcomingEvents} />
+        )}
+        {tabIndex === 1 && (
+          <EventList events={pastEvents} />
+        )}
     </Box>
-  );
-}
+      </>
+    );
+  }
